tasks/bulk-processing/requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
google-api-core==2.8.2
google-auth==2.8.0
google-cloud-core==2.3.1
google-cloud-pubsub==2.13.0
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
grpc-google-iam-v1==0.12.4
grpcio==1.45.0
grpcio-status==1.45.0
idna==3.3
proto-plus==1.20.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.0
rsa==4.8
six==1.16.0
urllib3==1.26.9
